Moved *IdlingResource classes from matchers to idlingresources package
